AU-2131: Base form class for shared logic, style fixes
