+++ public/modules/custom/grants_admin_applications/src/Form/ResendApplicationsForm.php
@@ -2,21 +2,18 @@
 
 namespace Drupal\grants_admin_applications\Form;
 
-use Drupal\Component\Serialization\Json;
 use Drupal\Core\Ajax\AjaxResponse;
 use Drupal\Core\Ajax\ReplaceCommand;
-use Drupal\Core\Form\FormBase;
 use Drupal\Core\Form\FormStateInterface;
 use Drupal\grants_handler\ApplicationHandler;
-use Drupal\helfi_atv\AtvDocument;
 use Drupal\helfi_atv\AtvService;
 use Symfony\Component\DependencyInjection\ContainerInterface;
 use Symfony\Component\HttpFoundation\Request;
 
 /**
- * Provides a grants_admin_applications form.
+ * Provides a grants_admin_resend_applications form.
  */
-class ResendApplicationsForm extends FormBase {
+class ResendApplicationsForm extends AtvFormBase {
 
   /**
    * Access to ATV.
@@ -45,7 +42,7 @@ class ResendApplicationsForm extends FormBase {
    * {@inheritdoc}
    */
   public function getFormId() {
-    return 'grants_admin_applications_admin_applications';
+    return 'grants_admin_applications_resend_applications';
   }
 
   /**
@@ -137,12 +134,11 @@ class ResendApplicationsForm extends FormBase {
     ];
 
     $res = \Drupal::service('helfi_atv.atv_service')->searchDocuments($sParams);
-    $atvDoc = reset($res);
-    return $atvDoc;
+    return reset($res);
   }
 
   /**
-   * GetStatus  submit handler.
+   * GetStatus submit handler.
    */
   public static function getStatus(array $form, FormStateInterface $formState) {
     $messenger = \Drupal::service('messenger');
@@ -217,75 +213,6 @@ class ResendApplicationsForm extends FormBase {
   }
 
   /**
-   * Attempts to resend ATV document through integrations.
-   *
-   * @param Drupal\helfi_atv\AtvDocument $atvDoc
-   *   The document to be resent.
-   * @param string $applicationId
-   *   Application id.
-   */
-  public static function sendApplicationToIntegrations(AtvDocument $atvDoc, string $applicationId) {
-    $httpClient = \Drupal::service('http_client');
-    $messenger = \Drupal::service('messenger');
-    $logger = self::getLoggerChannel();
-
-    $headers = [];
-    // Current environment as a header to be added to meta -fields.
-    $headers['X-hki-appEnv'] = ApplicationHandler::getAppEnv();
-    $headers['X-hki-applicationNumber'] = $applicationId;
-    $content = $atvDoc->getContent();
-    $myJSON = Json::encode($content);
-
-    // Usually we set drafts to submitted state before sending to integrations,
-    // should we do the same here?
-    $endpoint = getenv('AVUSTUS2_ENDPOINT');
-    $username = getenv('AVUSTUS2_USERNAME');
-    $password = getenv('AVUSTUS2_PASSWORD');
-
-    try {
-      $res = $httpClient->post($endpoint, [
-        'auth' => [
-          $username,
-          $password,
-          "Basic",
-        ],
-        'body' => $myJSON,
-        'headers' => $headers,
-      ]);
-
-      $status = $res->getStatusCode();
-
-      $messenger->addStatus('Integration status code: ' . $status);
-
-      $body = $res->getBody()->getContents();
-      $messenger->addStatus('Integration response: ' . $body);
-
-      $logger->info(
-        'Application resend - Integration status: @status - Response: @response',
-        [
-          '@status' => $status,
-          '@response' => $body,
-        ]
-      );
-    }
-    catch (\Exception $e) {
-      $logger->error('Application resending failed: @error', ['@error' => $e->getMessage()]);
-      $messenger->addError(t('Application resending failed: @error', ['@error' => $e->getMessage()]));
-    }
-  }
-
-  /**
-   * Gets the logger channel.
-   *
-   * @return \Psr\Log\LoggerInterface
-   *   The logger for the given channel.
-   */
-  public static function getLoggerChannel() {
-    $loggerFactory = \Drupal::service('logger.factory');
-    return $loggerFactory->get('grants_admin_applications');
-  }
-
-  /**
    * Ajax callback event.
    *
    * @param array $form
@@ -302,13 +229,8 @@ class ResendApplicationsForm extends FormBase {
    */
   public function ajaxCallback(array &$form, FormStateInterface $form_state, Request $request): AjaxResponse {
     $response = new AjaxResponse();
-    $response->addCommand(new ReplaceCommand('.grants-admin-applications-admin-applications', $form));
+    $response->addCommand(new ReplaceCommand('.grants-admin-applications-resend-applications', $form));
     return $response;
   }
-
-  /**
-   * {@inheritdoc}
-   */
-  public function submitForm(array &$form, FormStateInterface $form_state) {}
 
 }
